+++ lib/praxis/collection.rb
@@ -1,5 +1,5 @@
 module Praxis
- class Collection < Attributor::Collection
+  class Collection < Attributor::Collection
     include Types::MediaTypeCommon
 
     def self.of(type)
@@ -22,10 +22,23 @@ module Praxis
     def self.member_type(type=nil)
       unless type.nil?
         @member_type = type
+        @views = nil
         self.identifier(type.identifier + ';type=collection') unless type.identifier.nil?
       end
 
       @member_type
+    end
+
+    def self.views
+      @views ||= begin
+        @member_type.views.each_with_object(Hash.new) do |(name, view), hash|
+          hash[name] = Praxis::CollectionView.new(name, @member_type, view)
+        end
+      end
+    end
+
+    def self.domain_model
+      @member_type.domain_model
     end
 
   end