+++ lib/praxis/controller.rb
@@ -4,10 +4,10 @@ require 'active_support/all'
 module Praxis
   module Controller
     extend ::ActiveSupport::Concern
-    
+
     # A Controller always requires the callbacks
     include Praxis::Callbacks
-    
+
     included do
       attr_reader :request
       attr_accessor :response
@@ -22,7 +22,7 @@ module Praxis
         definition.controller = self
         Application.instance.controllers << self
       end
-      
+
       def id
         self.name.gsub('::'.freeze,'-'.freeze)
       end
@@ -36,5 +36,14 @@ module Praxis
     def definition
       self.class.definition
     end
+
+    def media_type
+      if (response_definition = self.request.action.responses[self.response.name])
+        return response_definition.media_type
+      else
+        self.definition.media_type
+      end
+    end
+
   end
 end